+++ backend/handlers/handlers_test.go
@@ -54,6 +54,7 @@ func (e *backendEnv) SystemOpen(string) error       { return nil }
 func (e *backendEnv) DeviceInfos() []usb.DeviceInfo { return nil }
 func (e *backendEnv) UsingMobileData() bool         { return false }
 func (e *backendEnv) NativeLocale() string          { return e.Locale }
+func (e *backendEnv) GetSaveFilename(string) string { return "" }
 
 func TestGetNativeLocale(t *testing.T) {
 	const ptLocale = "pt"